src/requirements.txt
numpy
torch
rasterio
tqdm
lightning
huggingface_hub
matplotlib
Pillow
scipy
configilm
hydra-core
ray
omegaconf
transformers
lmdb
psutil
pandas
safetensors
timm